+++ js/packages/cli/src/helpers/various-coinfra.ts
@@ -1,0 +1,23 @@
+import { CLUSTERS, DEFAULT_CLUSTER } from './constants';
+
+export function chunks(array, size) {
+  return Array.apply(0, new Array(Math.ceil(array.length / size))).map(
+    (_, index) => array.slice(index * size, (index + 1) * size),
+  );
+}
+
+export function parseDate(date) {
+  if (date === 'now') {
+    return Date.now() / 1000;
+  }
+  return Date.parse(date) / 1000;
+}
+
+export function getCluster(name: string): string {
+  for (const cluster of CLUSTERS) {
+    if (cluster.name === name) {
+      return cluster.url;
+    }
+  }
+  return DEFAULT_CLUSTER.url;
+}
